+++ packages/react-test-renderer/src/ReactTestRendererAct.js
@@ -15,7 +15,6 @@ import {
 } from 'react-reconciler/inline.test';
 import ReactSharedInternals from 'shared/ReactSharedInternals';
 import warningWithoutStack from 'shared/warningWithoutStack';
-import {warnAboutMissingMockScheduler} from 'shared/ReactFeatureFlags';
 import enqueueTask from 'shared/enqueueTask';
 import * as Scheduler from 'scheduler';
 
@@ -24,27 +23,11 @@ const {IsSomeRendererActing} = ReactSharedInternals;
 // this implementation should be exactly the same in
 // ReactTestUtilsAct.js, ReactTestRendererAct.js, createReactNoop.js
 
-let hasWarnedAboutMissingMockScheduler = false;
 const isSchedulerMocked =
   typeof Scheduler.unstable_flushAllWithoutAsserting === 'function';
 const flushWork =
   Scheduler.unstable_flushAllWithoutAsserting ||
   function() {
-    if (warnAboutMissingMockScheduler === true) {
-      if (hasWarnedAboutMissingMockScheduler === false) {
-        warningWithoutStack(
-          null,
-          'Starting from React v17, the "scheduler" module will need to be mocked ' +
-            'to guarantee consistent behaviour across tests and browsers. To fix this, add the following ' +
-            "to the top of your tests, or in your framework's global config file -\n\n" +
-            'As an example, for jest - \n' +
-            "jest.mock('scheduler', () => require.requireActual('scheduler/unstable_mock'));\n\n" +
-            'For more info, visit https://fb.me/react-mock-scheduler',
-        );
-        hasWarnedAboutMissingMockScheduler = true;
-      }
-    }
-
     let didFlushWork = false;
     while (flushPassiveEffects()) {
       didFlushWork = true;
