+++ packages/shared/forks/ReactFeatureFlags.native-fb.js
@@ -34,7 +34,7 @@ export const warnAboutDeprecatedSetNativeProps = true;
 export const enableFlareAPI = false;
 export const enableFundamentalAPI = false;
 export const enableJSXTransformAPI = false;
-export const warnAboutMissingMockScheduler = true;
+export const warnAboutUnmockedScheduler = true;
 export const revertPassiveEffectsChange = false;
 export const enableUserBlockingEvents = false;
 export const enableSuspenseCallback = false;
