+++ packages/shared/forks/ReactFeatureFlags.native-oss.js
@@ -29,7 +29,7 @@ export const warnAboutDeprecatedSetNativeProps = false;
 export const enableFlareAPI = false;
 export const enableFundamentalAPI = false;
 export const enableJSXTransformAPI = false;
-export const warnAboutMissingMockScheduler = false;
+export const warnAboutUnmockedScheduler = false;
 export const revertPassiveEffectsChange = false;
 export const enableUserBlockingEvents = false;
 export const enableSuspenseCallback = false;
