+++ packages/shared/forks/ReactFeatureFlags.persistent.js
@@ -29,7 +29,7 @@ export const warnAboutDeprecatedSetNativeProps = false;
 export const enableFlareAPI = false;
 export const enableFundamentalAPI = false;
 export const enableJSXTransformAPI = false;
-export const warnAboutMissingMockScheduler = true;
+export const warnAboutUnmockedScheduler = true;
 export const revertPassiveEffectsChange = false;
 export const enableUserBlockingEvents = false;
 export const enableSuspenseCallback = false;
