+++ packages/shared/forks/ReactFeatureFlags.test-renderer.www.js
@@ -30,7 +30,7 @@ export const disableJavaScriptURLs = false;
 export const enableFlareAPI = true;
 export const enableFundamentalAPI = false;
 export const enableJSXTransformAPI = true;
-export const warnAboutMissingMockScheduler = true;
+export const warnAboutUnmockedScheduler = true;
 export const enableUserBlockingEvents = false;
 export const enableSuspenseCallback = true;
 export const warnAboutDefaultPropsOnFunctionComponents = false;
